updated getNearestBank to support distance parameter

+++ AtmMoney/Engine/GetNearestBanks.h
@@ -17,7 +17,7 @@ typedef void (^VoidBlock)(void);
 @property (nonatomic, strong) NSMutableArray *banksData;
 @property (nonatomic, strong) NSMutableArray *bankHistoryData;
 
-- (void)getNearestBanksWithLocation:(CLLocation *)location andDistance:(NSInteger )distance withCompletion:(VoidBlock)completion andFailure:(VoidBlock)failure;
+- (void)getNearestBanksWithLocation:(CLLocation *)location andDistance:(CGFloat)distance withCompletion:(VoidBlock)completion andFailure:(VoidBlock)failure;
 
 - (void)getBankHistoryWithId:(NSInteger)buid withCompletion:(VoidBlock)completion andFailure:(VoidBlock)failure;
